(bmitem!) move current bmunit to bmitem
